Minor changes, key as flag

+++ main.go
@@ -1,15 +1,17 @@
 package main
 
 import (
+	"flag"
 	"net/http"
 
 	"github.com/gordonklaus/portaudio"
-	"github.com/joho/godotenv"
 )
 
 var Client *http.Client
 var Stop_Chan chan struct{}
 var Pipeline *PipeLine
+
+var OPENAI_KEY string
 
 func check_err(err error) {
 	if err != nil {
@@ -18,17 +20,20 @@ func check_err(err error) {
 }
 
 func Initialize_system() {
-	check_err(godotenv.Load())
-
 	Stop_Chan = make(chan struct{})
 	Client = &http.Client{}
 	Pipeline = getNewPipeLine()
 	portaudio.Initialize()
-
 }
 
 func main() {
 
+	flag.StringVar(&OPENAI_KEY, "OPENAI_KEY", "", "Provide Open AI Api key")
+	flag.Parse()
+
+	if OPENAI_KEY == "" {
+		panic("No API Key Provided")
+	}
 	Initialize_system()
 	defer portaudio.Terminate()
